absence-add-get

+++ Service-absence/src/main/java/io/absence/repository/AbsenceRepository.java
@@ -3,7 +3,7 @@ package io.absence.repository;
 import io.absence.entity.Absence;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AbsenceRepository extends JpaRepository<Absence, Integer> {
+public interface AbsenceRepository extends JpaRepository<Absence,Long> {
